add fix in lesson6
