+++ src/ru/itmo/java/basics/lesson6/UserSubUser.java
@@ -3,12 +3,16 @@ package ru.itmo.java.basics.lesson6;
 import java.util.Scanner;
 public class UserSubUser {
 
+
     static abstract class User {
-        public int getAge() {
+
+        public String getAge() {
             Scanner scanner = new Scanner(System.in);
             System.out.print("Введите ваш возраст: ");
-            return scanner.nextInt();
+            String age = scanner.nextLine();
+            return age;
         }
+
         public abstract String getName();
     }
 
@@ -17,19 +21,17 @@ public class UserSubUser {
         public String getName() {
             System.out.print("Введите ваше имя: ");
             Scanner scanner = new Scanner(System.in);
-            return scanner.nextLine();
+            String name = scanner.nextLine();
+            return name;
         }
     }
 
-    public static class Main {
-        public static void main(String[] args) {
-            SubUser user = new SubUser();
-            int age = user.getAge();
-            String name = user.getName();
-
-            System.out.println("Ваш возраст: " + age);
-            System.out.println("Ваше имя: " + name);
-        }
+    public static void main(String[] args) {
+        SubUser user = new SubUser();
+        String age = user.getAge();
+        String name = user.getName();
+        System.out.println("Ваш возраст: " + age);
+        System.out.println("Ваше имя: " + name);
     }
 }
